Update and re-arrange Gaussing Sampling code.
